Packaging cleanup and targeted community packages
This is 90% complete and includes previously reviewed and discussed cleanup, and
new targeted community packages based on Enterprise packages but without
Enterprise features. The repo layout is based on our Enterprise package repos.
It also includes using a new "testing" repo directory instead of directories
named for the release series, used as the target location for rc's and nightly
packages, which are not fully implemented yet. There's an artificial base
fileset in the code review with Enterprise spec files and packaging script
copied as base files for the new SSL files since that's the easiest way to
review just what's new about them compared to the existing packages.

We have options for how to differentiate these from other packages.  This
solution gives us new "mongodb-ssl" packages, distinct from Enterprise and
non-targeted Community packages. Alternative options are to name the packages
something else (like including "targeted" or similar wording in the names), or
using the same names as Community packages but in different repo locations. I
don't think we should reuse existing Community package names and would love a
better modifier than "-ssl" but I think that will be clearest to users.